improve dealing with the console when elevated
We can't directly connect the child to our console.
If the child does something directly with the console it will go
to the freshly allocated console that was force upon it.
We were hiding it, but that's a worse UX in that case than letting
it pop up or flicker open.

It can't be made perfect without some kind of pty proxy helper.