persuade windows that we really want utf-8 input/output
I'm skeptical that this will help, but it at least makes things
explicit; this commit adds a manifest that sets these utilities
into UTF-8 mode, and the bridge process will force the PTY console
into UTF-8 mode also.

That should make most of the pipeline UTF-8 everything.
However, we can't force the hosting console of `normdo` or `eledo`
into UTF-8 mode (well, we can, but we don't own it, so we shouldn't!).

This may help with refs: https://github.com/wez/deelevate/issues/1